requirements.txt
# PDF処理
PyMuPDF==1.24.1
Pillow==10.2.0

# Google Cloud APIs
google-cloud-vision==3.7.2
google-generativeai==0.7.2

# Notion API
notion-client==2.2.1

# ファイル監視
watchdog==4.0.0

# 設定・環境変数管理
PyYAML==6.0.1
python-dotenv==1.0.1

# HTTP リクエスト
requests==2.31.0

# PubMed検索
biopython==1.84

# 非同期処理
asyncio-throttle==1.0.2

# ログ設定
colorlog==6.8.2

# データ検証
pydantic==2.6.1

# 日付・時刻処理
python-dateutil==2.8.2